requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
kagglehub
numpy
kaggle
pymongo
python-dotenv